fix uglify and beefy commands
I guess the reason we used `uglifyjs` instead of `uglify-js` is because the former, according to its description, works with browserify. It seems that the repo has been deleted though; after testing the official uglify-js it seems to work with browserify now. The second reason is that the official one exposes the cli command; previously, running `npm run browserify-min` errors if no global installation of uglifyjs was found.

For the same reason, the small change to beefy is fine; local commands are given priority over global ones.